function register

+++ src/main/java/t221124nqt/ecommerce/hair_shop/controller/UserController.java
@@ -12,14 +12,14 @@ import t221124nqt.ecommerce.hair_shop.domain.response.user.ResCreateUserDTO;
 import t221124nqt.ecommerce.hair_shop.domain.response.user.ResGetUserDTO;
 import t221124nqt.ecommerce.hair_shop.domain.response.user.ResUpdateUserDTO;
 import t221124nqt.ecommerce.hair_shop.service.UserService;
-
-import javax.naming.Binding;
+import t221124nqt.ecommerce.hair_shop.util.anotation.ApiMessage;
+import t221124nqt.ecommerce.hair_shop.util.exception.EmailException;
+import t221124nqt.ecommerce.hair_shop.util.exception.IdInvalidException;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -38,12 +38,10 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public ResponseEntity<ResCreateUserDTO> createUser(@Valid @RequestBody User user, BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            System.out.println(bindingResult.getAllErrors());
-            return ResponseEntity
-                    .status(HttpStatus.BAD_REQUEST)
-                    .body(null);
+    @ApiMessage(message = "Tạo mới người dùng")
+    public ResponseEntity<ResCreateUserDTO> createUser(@Valid @RequestBody User user) throws EmailException {
+        if(this.userService.checkExistEmail(user.getEmail()) == true){
+            throw new EmailException("Email đã tồn tại!");
         }
         User newUser = this.userService.createUser(user);
         ResCreateUserDTO resCreateUserDTO = this.userService.convertToResCreateUserDTO(newUser);
@@ -53,27 +51,40 @@ public class UserController {
     }
 
     @GetMapping("/users")
+    @ApiMessage(message = "Lấy danh sách người dùng")
     public ResponseEntity<ResPaginationDTO> getUserById(@Filter Specification<User> spec, Pageable pageable) {
         ResPaginationDTO resPaginationDTO = this.userService.getAllUsers(spec, pageable);
         return ResponseEntity.ok().body(resPaginationDTO);
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<ResGetUserDTO> getUserById(@PathVariable("id") long id) {
+    @ApiMessage(message = "Tìm kiếm một người dùng")
+    public ResponseEntity<ResGetUserDTO> getUserById(@PathVariable("id") long id) throws IdInvalidException {
+        if(this.userService.checkExistId(id) == false){
+            throw new IdInvalidException("Không tồn tại User có id: " + id);
+        }
         User user = this.userService.getUserById(id);
         ResGetUserDTO resGetUserDTO = this.userService.convertToResGetUserDTO(user);
         return ResponseEntity.ok().body(resGetUserDTO);
     }
 
     @PutMapping("/users")
-    public ResponseEntity<ResUpdateUserDTO> updateUser(@Valid @RequestBody User user) {
+    @ApiMessage(message = "Cập nhật thông tin người dùng")
+    public ResponseEntity<ResUpdateUserDTO> updateUser(@Valid @RequestBody User user) throws IdInvalidException {
+        if(this.userService.checkExistId(user.getId()) == false){
+            throw new IdInvalidException("Không tồn tại User có id: " + user.getId());
+        }
         User currentUser = this.userService.updateUser(user);
         ResUpdateUserDTO resUpdateUserDTO = this.userService.convertToResUpdateUserDTO(currentUser);
         return ResponseEntity.ok().body(resUpdateUserDTO);
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) {
+    @ApiMessage(message = "Xóa người dùng")
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        if(this.userService.checkExistId(id) == false){
+            throw new IdInvalidException("Không tồn tại User có id: " + id);
+        }
         this.userService.deleteUser(id);
         return ResponseEntity.ok().body(null);
     }
